Transitioning from UIAlertView to UIAlertAction
In iOS8 UIAlertView was replaced by UIAlertAction, a new and improved API for displaying alerts that makes use of callbacks instead of delegation. This makes the code easier to read as all the relevant logic can be kept in one place and so we can see what would happen when the user interacts with the alert.

Since it is only a small change we have migrated before the old UIAlertViews are removed from iOS for good. Also increasing the deployment target to iOS 9 since we can no longer support older devices  and most devices are on iOS 9 or newer.